// src/app/appCommon/components/attachmnets/model/attachment-types.enum.ts
export enum AttachmentTypesEnum {
  PDF = "application/pdf",
  PNG_IMAGE ="image/png",
  JPEG_IMAGE = "image/jpeg",
  EXCEL = "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
  XLS = "application/vnd.ms-excel",
  DOCX = "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
  TXT = "text/plain",
  RICH_TEXT = "application/msword"
}